Update consent type check
